Separate out query only CL logic into its own file
